exercise1
(0) Establish the basic html with the HTML tag structure above
(1) Add a comment “Sep. 1, 2020” under <!DOCTYPE html>. 
(1) Set the title of your page “The page of your ID”
(2) Add an h1 heading on top of the page “Welcome to my page!”. And make the heading Italic.
(3) Add an h2 heading “My favorite sites:”. And add underline to the heading.
(4) Add a list for the following items (□ is a non-breaking space):
□□$Youtube®
□□$Wikipedia®
□□$Amazon®
□□$Facebook®
(5) Add following paragraphs in you page.
(CNN) If you're looking for a reason to care about tree loss, this summer's record-breaking heat waves
might be it. Trees can lower summer daytime temperatures by as much as 10 degrees Fahrenheit,
according to a recent study.
But tree cover in US cities is shrinking. A study published last year by the US Forest Service found that
we lost 36 million trees annually from urban and rural communities over a five-year period. That's a 1%
drop from 2009 to 2014.
(6) Align the first paragraph to the right. And align the second paragraph to the center.
(7) Add a blue horizontal line between the two paragraphs and make sure that its width is exactly 50% of
the screen. (color: #0000ff). Also align it to the left of the page.
(8) Change the font size of the first paragraph to 6. And change font color of the second paragraph to
orange. (color: #ff8000)
(9) Set the background of the entire page to gray. (color: #808080)