exercise 8 demand
(0) Import the js file. In the js file, one string is initialized and it should contain over 200 the most common male baby names in the UK.
(1) Split the string into an Array with all the names.
(2) Create a dropdown list contains all the names and give each name a sequential number.
(3) While pressing the select button, the page should create a pyramid with all the names, and the selected name should be highlighted with any color of your choice.
(4) Based on the current locale, add a line of the time information above the pyramid. Make sure the time will be updated while any updates are made on the page.
(5) Add an “I feel lucky!” button. While pressing the button, one of the names in the pyramid will randomly be highlighted.
(6) Add a “Let’s Disco!” button. While pressing the button, one of the names in the pyramid will randomly be highlighted every 1 second. (Hint: setInterval) And change the text on the “Let’s Disco!” button to “Stop the music!”.
(7) While pressing the “Stop the music!” button, the “Disco” should stop, and the text on the “Stop the music!” button should be changed back to “Let’s Disco!”.