exercise 9 demand
(0) Import the js file. The file contains a shoter list of the most common baby name in the U.K. Insert <table id="name_table"></table> under the body.
(1) Use a loop, getElementById(), createTextNode(), createElement(), and appendChild() to create a table with the names.
You may use HTML tags and innerHTML to finish the rest requirements of this note.
(2) Showing how many names are there in the table.
(3) Create a input area for users to input their first name and last name. And ask the users to separate the first name and last name by a space.
(4) Validate that there are only one first name and one last name in the input. If the input does not meet the requirements, show alert(“Wrong input”);
(5) If the name passed the validation of 4, Insert the first name into the table and update the total number of all names.
(6) Use setAttribute to set the border of the table = 1.
(7) When users copy any name from the table, show the following alert “NAME is copied!!” (Use addEventListener, https://developer.mozilla.org/en-US/docs/Web/Events, this.innerText).